requirements.txt
fastapi==0.111.0
uvicorn==0.30.1

pandas==2.2.2
numpy==1.26.4

beautifulsoup4==4.12.3
lxml==5.2.2

sentence-transformers==3.0.1
torch==2.3.1

faiss-cpu==1.8.0

pydantic==2.7.4
python-dotenv==1.0.1
pyyaml==6.0.1

ruff==0.5.1
black==24.4.2

pytest==8.2.2
google-genai